refactor(dmworkdrive): move IM transferred-state wire types into bridge/types
Address the PR #1261 blocking review comment: the two new wire contracts
(`ImTransferredEntry`, `ImTransferredItem`) lived in `api/driveApi.ts`, not in
`bridge/types.ts`, whose header explicitly states it is the file that mirrors
octo-drive Go DTOs 1:1. The saved-state detection rests on the shared
inter-service source_key format
`${im_channel_type}#${im_group_no}#${im_msg_id}`; pinning it in bridge/types
makes the contract discoverable next to the other backend DTO mirrors and
documents the delimiter/format so a future backend or frontend change cannot
silently drift the two keys apart (which was the reviewer's silent-failure-mode
concern — misaligned keys would resolve `null` for every card and show every
already-saved file as "Save to Drive" forever, with no error/warning).

- bridge/types.ts: add `ImTransferredEntry` / `ImTransferredItem` with a large
  header comment fixing the source_key format, delimiter choice ('#' cannot
  collide with hex uid / peer uid / "group_no____short_id" sub-thread ids), the
  meaning of every field, and the paired backend surface (drive_file.source_key
  column + POST /blobs/im-transferred/batch + migration 007).
- api/driveApi.ts: drop the local declarations, re-export the two types from
  bridge/types so existing importers (module.tsx, App.tsx, tests) keep working
  unchanged.

No behavior change; wire is byte-identical. tsc / i18n:check unchanged from
baseline (dmworkdrive 142 errors == baseline 142; dmworkbase Messages/File
30 errors == baseline 30 — all are pre-existing module-resolution noise when
tsc is run per-package).

Pairs with octo-server PR #708 (DM message get + Space isolation) and
octo-drive MR !5 (source_key column + Person routing).
